+++ lib/rubocop-rspec.rb
@@ -9,21 +9,23 @@ require_relative 'rubocop/rspec'
 require_relative 'rubocop/rspec/version'
 require_relative 'rubocop/rspec/inject'
 require_relative 'rubocop/rspec/node'
-require_relative 'rubocop/rspec/top_level_describe'
 require_relative 'rubocop/rspec/wording'
 require_relative 'rubocop/rspec/language'
 require_relative 'rubocop/rspec/language/node_pattern'
-require_relative 'rubocop/rspec/top_level_group'
+
+require_relative 'rubocop/cop/rspec/mixin/top_level_describe'
+require_relative 'rubocop/cop/rspec/mixin/top_level_group'
+require_relative 'rubocop/cop/rspec/mixin/variable'
+require_relative 'rubocop/cop/rspec/mixin/final_end_location'
+require_relative 'rubocop/cop/rspec/mixin/empty_line_separation'
+
 require_relative 'rubocop/rspec/concept'
 require_relative 'rubocop/rspec/example_group'
 require_relative 'rubocop/rspec/example'
 require_relative 'rubocop/rspec/hook'
-require_relative 'rubocop/rspec/variable'
 require_relative 'rubocop/cop/rspec/cop'
 require_relative 'rubocop/rspec/align_let_brace'
 require_relative 'rubocop/rspec/factory_bot'
-require_relative 'rubocop/rspec/final_end_location'
-require_relative 'rubocop/rspec/empty_line_separation'
 require_relative 'rubocop/rspec/corrector/move_node'
 
 RuboCop::RSpec::Inject.defaults!
