+++ lib/rubocop/cop/rspec/describe_class.rb
@@ -22,7 +22,7 @@ module RuboCop
       #   describe "A feature example", type: :feature do
       #   end
       class DescribeClass < Cop
-        include RuboCop::RSpec::TopLevelDescribe
+        include TopLevelDescribe
 
         MSG = 'The first argument to describe should be '\
               'the class or module being tested.'
