+++ lib/rubocop/cop/rspec/multiple_describes.rb
@@ -23,7 +23,7 @@ module RuboCop
       #     end
       #   end
       class MultipleDescribes < Cop
-        include RuboCop::RSpec::TopLevelDescribe
+        include TopLevelDescribe
 
         MSG = 'Do not use multiple top level describes - '\
               'try to nest them.'
